requirements.txt
streamlit
psycopg2-binary
pandas
pyyaml
python-dotenv
requests
schedule
SQLAlchemy